Fix CXX_BUILD.

+++ gfx/context/glx_ctx.c
@@ -379,8 +379,8 @@ static bool gfx_ctx_set_video_mode(
       if (g_core)
       {
          const int attribs[] = {
-            GLX_CONTEXT_MAJOR_VERSION_ARB, g_major,
-            GLX_CONTEXT_MINOR_VERSION_ARB, g_minor,
+            GLX_CONTEXT_MAJOR_VERSION_ARB, (int)g_major,
+            GLX_CONTEXT_MINOR_VERSION_ARB, (int)g_minor,
             GLX_CONTEXT_PROFILE_MASK_ARB, GLX_CONTEXT_CORE_PROFILE_BIT_ARB,
             GLX_CONTEXT_FLAGS_ARB, g_extern.system.hw_render_callback.debug_context ? GLX_CONTEXT_DEBUG_BIT_ARB : 0,
             None,
